refactor(userPage): separate mock files
