+++ src/pages/privileges/outlets/users/config/invitationsTable.config.jsx
